services/orchestrator/pyproject.toml
[project]
name = "nexus-orchestrator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.4.0",
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
    "protobuf>=5.0",
    "google-cloud-pubsub>=2.20",
    "google-cloud-firestore>=2.16",
    "redis>=5.0",
    "httpx>=0.27",
    "pydantic>=2.6",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
